split BaseAbstractAlgebra into multiple files
